requirements.txt
pytest
pyyaml
jsonschema
aioconsul
